Update all.pb to match .proto file
The `all.pb` file was generated from an older version of `echo.proto`
which still had a version number in the package. It was not updated
during review of PR #2 when the version number was removed from the
package name.

Updated by running:

    make pb

Update the Makefile so that `check-uptodate` builds `all.pb` so CI will
fail in future if it is not updated when it should be.