Initial version of `pony serve`
Add the initial mvp implementation of `pony serve` that can use jsonnet
to serve unary RPCs. Client and/or server streaming is not yet
implemented and will result in an "Unimplemented" gRPC status result.

Adjust the test client and server default port to 8080 to align with
pony's default listen port. 8080 is more standard for unencrypted HTTP
traffic.