Add all streaming to echo service and samples
Extend the echo service to have four methods so that all combinations of
unary/streaming client/server are there. Update the client and server to
work with all these methods, and remove the dynamic server as it is no
longer needed now that pony serve does dynamic grpc.

Protos stuff was all updated with:

    make protos pb