👗 Handle status responses from jsonnet methods (#6)
Handle the `status` field in a jsonnet method response as a Status
protobuf and return that as an error when present.

Update the sample client to output status details when printing a gRPC
error.

Update the unary echo service jsonnet method definition to return a
status instead of a response when the message is "Bart", and include a
sample `details` field.

Update the README to document the status result of a method.

Pull-Request: https://github.com/foxygoat/jig/pull/6