client: Output header/trailer metadata
Output the header and trailer metadata for each of the four types of
method calls it makes. Ensure it is output even when an error is
returned as metadata accompanies both error and non-error results.

Update the bidi streaming method to concurrently run the send and
receive streams as it was incorrect before with how jig works - one
message on the send stream can result in multiple messages on the
receive stream coming back and we were just expecting one. We were also
not properly handling EOF on the receive stream.