Rename BiDi to Bidi
Rename all occurrences of the abbreviation "BiDi" to "Bidi" (lowercase
d). It is an abbreviation of a single word (bidirectional) not two
words, so the correct camel casing is "Bidi".

This also renames a couple of files as there is a gRPC method named with
"Bidi" in it and the jsonnet files are named after the method.

Mac users (any any others with case-insensitive filenames) may need to
double check the filenames:

    testdata/greet.Greeter.HelloBidiStream.jsonnet
    serve/testdata/sample/greet.Greeter.HelloBidiStream.jsonnet

and ensure they have the right case. It may be necessary to remove the
files and run `git restore .` to get them back with the right name.