bones: Add sub-command to generate jsonnet method skeletons
Add the `bones` subcommand to `jig` to generate skeleton jsonnet files
for the methods in a protoset. The skeleton includes the zero value for
all basic types, a message has all fields populated with the zero value,
and a repeated field (list) has one element. This should make it easy to
see the structure of a methods input and output messages.

The subcommand can limit what it creates with an optional list of names
which are flexibly interpreted as either a full method name
(pkg.service.method), a full service name (pkg.service), a package name,
a bare service name or a bare method name. For packages and services,
all methods in it are generated.

By default, existing files will not be overwritten unless `--force/-f`
is specified. This allows you to easily generate skeletons for new
methods without overwriting any existing jsonnet method definitions.

If a `--method-dir/-m` is not specified, the skeletons are written to
standard out.

Skeletons do not contain any extension fields as yet.