proto: Generate per-proto .pb files
Update the `proto` target (was called `protos` inconsistently) to
generate `.pb` protoset files as well as Go code. This allows more
targeted use of jig to work/test against a single service in a `.pb`
file rather than all of them.

Remove the `pb` make target as per-proto `.pb` files should be
sufficient for everything and are clearer about just what is in them.
Update the tests to use `greeter.pb` instead of `all.pb`, and remove
`all.pb` from the repo.

Sort the `find` output that finds all the `.proto` files so they are
processed in a deterministic order. This used to matter with the
`all.pb` file, but even with that gone, it is still nice to have things
done in the same order when building.

`greeter.pb` generated by running:

    make proto