Add grpc server reflection
Add an implementation of the `ServerReflection` gRPC service, reflecting
over the `protoregistry.Files` registry used by the `serve` command.

This re-implements the service instead of using the existing
implementation as the existing implementation reflects over the global
registry which is not populated for our use case.

Add a wrapper for the `protoregistry.Files` type to implement the
`ExensionTypeResolver` and `MessageTypeResolver` so that it can be used
as a resolver for marshaling and unmarshaling text formats. The
`ExtensionTypeResolver` is useful for implementing one of the
ServerReflection requests (FileContainingExtensions).
`MessageTypeResolver` is currently unimplemented (TODO) but will be
implemented for completeness.

Link: https://github.com/grpc/grpc/blob/master/src/proto/grpc/reflection/v1alpha/reflection.proto