⚡ registry: Remove unnecessary loop+allocation from walk (#18)
Have the `range*` functions return `protoreflect.ExtensionType` directly
instead of `protoreflect.ExtensionDescriptor` so that we do not need a
loop + slice allocation at the end of `walkExtensions` to convert
`ExtensionDescriptor` into `ExtensionType`. The latter is only ever used
to convert to the former, so just generate that in the first place.

Pull-Request: https://github.com/foxygoat/jig/pull/18