🍦 Changer serve.Server to align with http.Server (#19)
Changer serve.Server to align with http.Server, implementing a Server
and ListenAndServe method. This makes it also easier to create
TestServer similar to httptest.Server.

This refactor was motivated by the desire to run a _ready_ jigServer in
a go routine as

   jigServer.Run()

or similar and to avoid separate network probing and readiness check.

Pull-Request: https://github.com/foxygoat/jig/pull/19