🌱 Introduce serve.WithFS() option (#21)
Introduce serve.WithFS() option so that we can use serve with embed.FS.
In order to make test embed.FS we cannot use links and we need to place
the .pb file in the same directory as the methods.

This is a temporary state until we cleanup NewServer in a more
comprehensive, not allowing for .pb files to be specified separately
from the method dirs.

Pull-Request: https://github.com/foxygoat/jig/pull/21