bones: Emit well-formed bones for well-known types
Render protobuf well-known types properly in JSON. Some well-known
protobuf types have a JSON form different to how their proto
representation would normally be rendered. For instance, timestamp and
duration are rendered in JSON as a single string, but are in the proto
file as a message. Boxed types are rendered as their value only and not
nested in a message.

Extend the test cases for all the well-known types by adding a new rpc
and response message for the well-known cases. Slightly modify the test
so the output is better when the golden file does not match.

Add regenerated proto output files for `exemplar.proto`.