registry: Simplify descriptor walking for extensions
Simplify the code for walking extensions in a descriptor tree by
introducing the `extensionContainer` interface which is implemented by
both FileDescriptor and MessageDescriptor. These are both namespaces
that can contain extensions and have the same interface for navigating
the namespace and getting extensions.

Redo `FindExtensionByName` to use the `files.FindDescriptorByName`
method as it works of an index rather than walking the descriptor tree.
This should be more efficient.