serve: Fix cli handling of protoset files
Pass protoset files specified on the command line with `--protoset` to
`serve.Server` as an option (`WithProtosets`). This got missed in a
recent refactor of method dir handling.

Add a couple of debug logs to see protosets being loaded - this is
generally useful as you don't always know you have the protset files
correct - this helps diagnose issues with protoset files.