serve: Remove method struct
Remove the `method` struct and move the methods to `Server`. Most of the
method struct (2 of the 3 fields) are the same for all methods and come
from `Server`. The other (the method descriptor) can just be passed into
the `callMethod` method.

This is being done as soon a new field would need to be added to
`method` that again would be the same for all methods.