serve: Use a Resolver when un/marshaling json
Use a `registry.Files` as a `Resolver` when mashaling and unmarshaling
json so that extensions are properly handled. This allows a jig method
implementation to accept and respond with messages containing extensions
with the json syntax of `"[pkg.field]"` for the extension field name.

Add a message with an extension field to the unary Hello greeter
response to demonstrate the use of extensions. This just uses the `http`
extension field of `MethodOptions` as we are already using that - no
need to define another message extension.

Vendor the duration.proto file so we can add the `duration.pb` file to
the greeter test directory as it is now needed for the current test case
of returning a status detail with a duration. This used to work as
`google.golang.org/grpc` linked with a few proto types for its binarylog
implementation meaning `duration.pb` was incidentally in the global
registry. Now we are not using the global registry, we need to
explicitly include `duration.pb` in the greet directory. This is copied
there in the Makefile.

Generate `duration.pb` with

    make proto

Remove the top-level testdata directory as it is unused itself and only
existed as a sample referenced by the README. Update the README to point
at `serve/testdata/greet` instead. This eliminates the duplication of
these files and also makes it simpler to start the server with the
example as the new directory also contains the `.pb` files needed.

Link: https://github.com/protocolbuffers/protobuf/blob/master/src/google/protobuf/duration.proto