serve: Remove unnecessary stream interceptor
Remove the stream interceptor in serve.Server and use just the unknown
handler. The interceptor was originally used to get the method being
called, however that is also available via the grpc.ServerStream passed
to the unknown handler by getting the grpc.ServerTransportStream from
the server stream context. The server transport stream as a `Method()`
method that returns the method being invoked.

This simplifies the implementation back to what was originally
envisaged, a dynamic server implementing the "unknown" methods.

The UnknownHandler is now exported as it also provides an entry point to
the jig server for other protocols (such as http), which will allow jig
to be easily re-used from outside the serve package.