serve: Use Server.UnknownHandler from httprule
Extend the `UnknownHandler` in `serve.Server` to optionally take a full
method name as the first `srv` parameter and have httprule.Server call
that instead of the internal `callMethod`. This will allow the
`httprule.Server` to be decoupled from `serve.Server` allowing other
implementations to be plugged in.

The `UnknownHandler` is changed to slightly abuse the first `srv`
parameter (an `interface{}`). For an grpc unknown handler, this is
always the `nil` value. For concrete implementations of
`grpc.StreamHandler`, this is the server instance that will handle the
grpc method. Since the unknown handler has no known implementation, only
`nil` can be passed. To pass a method name, it should be of type
`protoreflect.FullName`.

This change makes it easier for non-grpc implementations to plug in as
otherwise they would need to provide an implementation of
`grpc.ServerTransportStream`, put it in the context accessible via
`grpc.ServerStream.Context()` and implement the `Method()` method to
return the method name. It would also need to implement three other
methods which are of no use.