serve: Decouple httprule from serve
Decouple `httprule.Server` from `serve.Server` by allowing any
`http.Handler` to be plugged into `serve.Server`. When a hander is
supplied, it causes the experimental `ServeHTTP` path of `grpc.Server`
to be used with golang.org/x/net/http2{,/h2c} instead of using the http2
server built into the grpc library.

Split the tests so those that depend on httprule are now in that
package, and add a simple test for a basic `http.Handler` to the serve
package.

Update main.go to take -h/--http to enable the httprule handler in the
jig cli.