reflection: Use foxygo.at/protog/registry
Use the files registry in `foxygo.at/protog/registry` instead of the one
in this repo. The one in protog comes from here and is where that
package will be hosted in future, removing it from this repo.

The use of the registry in the reflection server is purely internal, so
migrate that first. Introduce a `FileDescriptorRanger` interface so that
the reflection server can be constructed with either a standard
`protoregistry.Files` instance or a `registry.Files` instance (or any
other collection of `FileDescriptor`s that can be ranged over).