serve: Use foxygo.at/protog/registry
Use the files registry in `foxygo.at/protog/registry` instead of the one
in this repo. The one in protog comes from here and is where that
package will be hosted in future, removing it from this repo.

As the new package is interface compatible with the old one for our
purposes, we only need to change the import. All other code stays the
same.