💌 Infer Content-Type and Accept header (#40)
Infer Content-Type and Accept header when not set: If Content-Type is
not set but Accept header is set, set the inferred content type as
fallback to the accept type. If the Accept header is not set but the
Content-Type header is set, infer the accept type from the
Content-Type.

This fix is proposed to more closely mirror CashApp internal custom
proto over HTTP/1.1 implementations.

Pull-Request: https://github.com/foxygoat/jig/pull/40