☠️ Fix bones for recursive message types (#42)
Fix `jig bones` for recursive message types so that it doesn't cause an
infinite recursion:

	runtime: goroutine stack exceeds 1000000000-byte limit

Add a stack to the formatter tracking the the message names. Rename
FormatOptions to Formatter to better reflect the carried state and type
alias it back to FormatOptions for compatibility and external use.

Pull-Request: https://github.com/foxygoat/jig/pull/42