httprule: Add optional "default" handler
Add an optional "default" handler to `httprule.Handler` that is called if
the request does not match any of the gRPC methods. The default is still
to return a 404 Not Found status if no "default" handler is supplied.

This allows chaining handlers so that other non-gRPC paths can be
handled by other handlers, with precedence to gRPC paths.