bones: Emit proto map as object not array of pairs [2/3]
Emit a proto `map<key, val>` as a JSON object instead of an array of
key/value pairs, with keys converted to strings. The array is the
internal proto representation of a map but the protojson format
specifies an object as the representation of a map. The bones exemplars
are not parsable by protojson, producing the error:

    proto: syntax error (line X:XX): unexpected token [.

It is possible that the array of pairs representation was parsable in
the past, but it is not true today and there is no support for schema
evolution for JSON at the same level as the proto wire format.

Gen-command: 2/3