test: Refactor server_test.go
Refactor server_test.go to be independent from internal/client package and less
brittle by avoiding sting comparison in favour of comparing map values for
certain keys for header and trailers. In a follow up commit we will upgrade
protoc-gen-go and libraries which made these tests fail, so we are fixing them
first and independently.